versioning + error/warn management + custom smoke tab

+++ action.js
@@ -31,13 +31,36 @@ const getActions = async () => {
     if (!!serverName) {
       out.group = serverName
     }
+    const paramsPath = `${__dirname}/actions/${action.key}/params.txt`
+    if (fs.existsSync(paramsPath)) {
+      const paramsFile = fs.readFileSync(paramsPath, 'utf8')
+      out.params = paramsFile.split(/\r?\n/)
+    }
     return out
   }) : []
   const botActions = await botManager.getActions()
   return myActions.length > 0 ? myActions.concat(botActions) : botActions
 }
 
-const runScript = (action, path) => {
+const afterScript = (code, logFile, logName, action, path, param, lock) => {
+  const result = code === 0 ? 'succeded' : code === 10 ? 'succeded, re-run requested' : code === 3 ? 'succeded with warnings' : 'failed'
+  log(logFile, `==== The "${action.name}" action script ${result} ====`)
+  const newLog = `${logName}-${code === null ? 1 : code}.log`
+  fs.renameSync(logFile, newLog)
+  if (code === 10) {
+    runScript(action, path, param)
+  } else {
+    fs.unlinkSync(lock)
+    notifierService.notifyAction(code === 0 ? 'ok' : code === 3 ? 'warn' : 'fail', action, newLog)
+  }
+}
+
+const errorHandler = (err, logFile) => {
+  console.error(err)
+  log(logFile, `== Error == \r\n${err.name}: ${err.message}`)
+}
+
+const runScript = async (action, path, param) => {
   const lock = `${path}/.lock`
   const scriptFile = `${path}/run.sh`
   const logName = `${path}/logs/${(new Date()).getTime()}`
@@ -46,42 +69,63 @@ const runScript = (action, path) => {
     log(logFile, `The "${action.name}" action script is starting`)
     const isWin = os.platform() === "win32"
     const spawnParams = isWin ? { shell: true } : {}
+    let hasWarning = false
     if (action.root) {
       spawnParams.cwd = action.root
     }
-    const process = spawn(scriptFile, spawnParams)
+    const process = spawn(scriptFile, param ? [param] : [], spawnParams)
     process.stdout.on('data', data => {
       log(logFile, data)
     })
     process.stderr.on('data', data => {
-      log(logFile, `== Stderr == \r\n${data}`)
-    })
-    process.on('close', code => {
-      log(logFile, `==== The "${action.name}" action script ${code === 0 ? 'succeded' : code === 10 ? 'succeded, re-run requested' : 'failed'} ====`)
-      const newLog = `${logName}-${code === null ? 1 : code}.log`
-      fs.renameSync(logFile, newLog)
-      if (code === 10) {
-        runScript(action, path)
-      } else {
-        fs.unlinkSync(lock)
-        notifierService.notifyAction('ok', action, newLog)
+      const errorRx = /error|fatal|failed|exception|critical/i
+      const warnRx = /warning|deprecated|caution|notice/i
+      let title = 'Stderr output'
+      if (errorRx.test(data)) {
+        title = 'Error'
+        process.exitCode = 1
+      } else if (warnRx.test(data)) {
+        title = 'Warning'
+        hasWarning = true
+      }
+      log(logFile, `== ${title} == \r\n${data}`)
+      if (process.exitCode === 1) {
+        process.kill()
       }
     })
+    process.on('uncaughtException', (err) => {
+      errorHandler(err, logFile, logName, lock, action)
+      process.exitCode = 1
+      process.kill()
+    })
+    process.on('unhandledRejection', (reason, promise) => {
+      errorHandler(err, logFile, logName, lock, action)
+      process.exitCode = 1
+      process.kill()
+    })
+    process.on('close', code => {
+      afterScript(hasWarning ? 3 : code, logFile, logName, action, path, param, lock)
+    })
     process.on('error', err => {
-      log(logFile, `error ${err.name}: ${err.message}`)
-      const newLog = `${logName}-1.log`
-      fs.renameSync(logFile, newLog)
-      fs.unlinkSync(lock)
-      notifierService.notifyAction('fail', action, newLog)
+      errorHandler(err, logFile, logName, lock, action)
+      process.exitCode = 1
+      process.kill()
     })
   } else {
-    log(logFile, `No ${action.name} script was found`)
-    fs.renameSync(logFile, `${logName}-0.log`)
-    fs.unlinkSync(lock)
+    const nodeFile = `${path}/run.js`
+    if (fs.existsSync(nodeFile)) {
+      const run = require(nodeFile)
+      const code = await run(log, logFile, param ? [param] : [])
+      afterScript(code, logFile, logName, action, path, param, lock)
+    } else {
+      log(logFile, `No ${action.name} script was found`)
+      fs.renameSync(logFile, `${logName}-0.log`)
+      fs.unlinkSync(lock)
+    }
   }
 }
 
-const execute = (actionKey, tries = 3) => {
+const execute = (actionKey, param, tries = 3) => {
   const action = actions.find(action => action.key === actionKey)
   if (action) {
     const path = `${__dirname}/actions/${action.key}`
@@ -96,7 +140,7 @@ const execute = (actionKey, tries = 3) => {
       }
     } else {
       fs.closeSync(fs.openSync(lock, 'w'))
-      runScript(action, path)
+      runScript(action, path, param)
     }
   } else {
     console.log(`No action found with the key "${actionKey}"`)
